move StyledComponents of core components to styles.tsx file
