+++ components/core/BlockText/BlockText.tsx
@@ -1,15 +1,11 @@
 import React from 'react';
-import styled from 'styled-components';
+
+import { StyledBlockText } from './styles';
 
 type BlockTextProps = {
   className?: string;
   text: string;
 };
-
-const StyledBlockText = styled.p`
-  color: ${({ theme }) => theme.palette.DARK_MEDIUM};
-  line-height: 1.2rem;
-`;
 
 export const BlockText: React.FC<BlockTextProps> = ({ className, text }) => (
   <StyledBlockText className={className}>{text}</StyledBlockText>
