+++ components/core/BlockText/styles.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const StyledBlockText = styled.p`
+  color: ${({ theme }) => theme.palette.DARK_MEDIUM};
+  line-height: 1.2rem;
+`;
