+++ components/core/Card/__stories__/card.story.tsx
@@ -3,9 +3,11 @@ import { storiesOf } from '@storybook/react';
 
 import { text } from '@storybook/addon-knobs';
 
-import { Card, CardType } from '../Card';
+import { Card } from '../Card';
 
 import { CARD_PROPS_MOCKS } from '../__mocks__/card.mocks';
+
+import { CardType } from '../../../../helpers/types';
 
 const stories = storiesOf('Card', module);
