+++ components/core/Card/__tests__/card.spec.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 
-import { Card, CardType } from '../Card';
+import { Card } from '../Card';
 
 import { CARD_PROPS_MOCKS } from '../__mocks__/card.mocks';
 
 import { renderWithTheme } from '../../../../helpers/Theme';
 import { CDN_URL_STATIC_DIRECTORY } from '../../../../helpers/utils';
+import { CardType } from '../../../../helpers/types';
 
 it('should render standart card', () => {
   const { container } = render(
