+++ components/core/Rating/styles.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const StyledRankStarsWrapper = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+export const StyledRankStar = styled.i`
+  font-size: 1.5rem;
+  color: ${({ theme }) => theme.palette.DARK_MEDIUM};
+`;
+
+export const StyledText = styled.p`
+  color: ${({ theme }) => theme.palette.DARK_MIN};
+  margin-left: 10px;
+`;
