+++ helpers/types.tsx
@@ -7,9 +7,20 @@ export type InitialAppState = {
 };
 
 // Core / ui components
+
+export enum LogoSize {
+  BIG = 'big',
+  SMALL = 'small',
+}
+
 export enum LoaderType {
   CIRCLE = 'circle',
   LINE = 'line',
+}
+
+export enum LoaderMode {
+  DARK = 'dark',
+  LIGHT = 'light',
 }
 
 export type ListItem = {
@@ -19,7 +30,19 @@ export type ListItem = {
   path: string;
 };
 
+export enum CardType {
+  SUGGESTION = 'suggestion',
+  STANDART = 'standart',
+  HIGHLIGHT = 'highlight',
+}
+
+export enum CustomLinkSize {
+  BIG = 'big',
+  SMALL = 'small',
+}
+
 // Restaurant data
+
 export type Timming = {
   id: string;
   day: string;
@@ -68,6 +91,7 @@ export type RestaurantDetail = {
 };
 
 // Breadcrumbs
+
 export enum BreadcrumbsType {
   HOME = 'home',
   SEARCH = 'search',
@@ -87,6 +111,7 @@ export type BreadcrumbsData = {
 };
 
 // Actions
+
 type AddFavoriteAction = {
   type: 'ADD_FAVORITE';
   favorite: Restaurant;
